[stdlib] fix return type of getNumRuntimeFunctionCounters
The return type of getNumRuntimeFunctionCounters is defined as
uint64_t in RuntimeInvocationsTracking.cpp, but it has return type
Int in RuntimeFunctionCounters.swift.

Found when compiling the stdlib for WebAssembly, as WebAssembly validates
return types. uint64_t corresponds to i64, but Int is i32,
so the program fails validation.

+++ stdlib/public/core/RuntimeFunctionCounters.swift
@@ -140,7 +140,7 @@ struct _RuntimeFunctionCounters {
   /// Get the number of different runtime functions whose calls are being
   /// tracked.
   @_silgen_name("_swift_getNumRuntimeFunctionCounters")
-  public static func getNumRuntimeFunctionCounters() -> Int
+  public static func getNumRuntimeFunctionCounters() -> UInt64
 
   /// Dump all per-object runtime function counters.
   @_silgen_name("_swift_dumpObjectsRuntimeFunctionPointers")
